+++ Toolkit/Tests/e2e/Get-RscSnapshot.Tests.ps1
@@ -11,7 +11,12 @@ BeforeAll {
 Describe -Name 'Get-RscSnapshot Tests' -Tag 'Public' -Fixture {
 
     It -Name 'retrieves Snapshots' -Test {
-        $data.snapshots = Get-RscVmwareVm | Select-Object -first 1 | Get-RscSnapshot
+        $data.snapshots = Get-RscVmwareVm -Relic:$false -Replica:$false | Select-Object -first 1 | Get-RscSnapshot
+        $data.snapshots | Should -Not -BeNullOrEmpty
+    }
+
+    It -Name 'retrieves Database Snapshots' -Test {
+        $data.snapshots = Get-RscMssqlDatabase -Relic:$false -Replica:$false | Select-Object -first 1 | Get-RscSnapshot
         $data.snapshots | Should -Not -BeNullOrEmpty
     }